refactor create user document methods
