+++ src/components/SignUp/SignUp.tsx
@@ -1,10 +1,4 @@
-import {
-  useState,
-  useCallback,
-  useContext,
-  FormEvent,
-  ChangeEvent,
-} from 'react';
+import { useState, useCallback, FormEvent, ChangeEvent } from 'react';
 import styled from 'styled-components';
 
 import { FirebaseError } from 'firebase/app';
@@ -16,7 +10,6 @@ import {
 
 import FormInput from '../FormInput/FormInput';
 import Button from '../Button/Button';
-import UserContext from '../../contexts/user';
 
 const INITIAL_FORM_FIELDS = {
   displayName: '',
@@ -27,8 +20,6 @@ const INITIAL_FORM_FIELDS = {
 const SignUp: React.FC = () => {
   const [formFields, setFormFields] = useState(INITIAL_FORM_FIELDS);
   const { displayName, email, password } = formFields;
-
-  const { setCurrentUser } = useContext(UserContext);
 
   const resetForm = () => {
     setFormFields(INITIAL_FORM_FIELDS);
@@ -65,8 +56,6 @@ const SignUp: React.FC = () => {
           email: user.email || '',
         };
 
-        setCurrentUser(userAuth);
-
         await createUserDocument(userAuth, { displayName });
 
         resetForm();
@@ -84,7 +73,7 @@ const SignUp: React.FC = () => {
         }
       }
     },
-    [email, password, displayName, setCurrentUser]
+    [email, password, displayName]
   );
 
   return (
